CFloat -> CDouble everywhere

+++ functions/CFunctionMV.cpp
@@ -4,7 +4,7 @@
 
 #include "CFunctionMV.h"
 
-CFunctionMV::CFunctionMV(string name, float f(float, float), float f_derivative_x(float, float), float f_derivative_y(float, float)) {
+CFunctionMV::CFunctionMV(string name, double f(double, double), double f_derivative_x(double, double), double f_derivative_y(double, double)) {
     this->function_definition = std::move(name);
     this->f = f;
     this->f_derivative_x = f_derivative_x;
