+++ functions/CFunctionSV.cpp
@@ -5,7 +5,7 @@
 #include "CFunctionSV.h"
 
 
-CFunctionSV::CFunctionSV(string name, string phi_name, float f(float), float f_derivative(float), float f_second_derivative(float), float phi(float), float phi_derivative(float)) {
+CFunctionSV::CFunctionSV(string name, string phi_name, double f(double), double f_derivative(double), double f_second_derivative(double), double phi(double), double phi_derivative(double)) {
     this->f = f;
     this->function_definition = std::move(name);
     this->function_definition_phi = std::move(phi_name);
